门牌号逻辑已完成 首页、分类已添加url，交互作业中(一级页面功能OK) 充值以及微信支付待做(充值待调用接口) 优化刷新加载的设计 修复了没有优惠券  显示无优惠券页面的bug

+++ app/src/main/java/com/jtsoft/letmedo/bean/BillOrderBean.java
@@ -160,7 +160,7 @@ public class BillOrderBean {
             private int userId;
             private double orderPrice;
             private double orderOldPrice;
-            private int expressPrice;
+            private double expressPrice;
             private Object couponReceiveId;
             private double couponPrice;
             private Object payTime;
@@ -230,7 +230,7 @@ public class BillOrderBean {
                 this.orderOldPrice = orderOldPrice;
             }
 
-            public int getExpressPrice() {
+            public double getExpressPrice() {
                 return expressPrice;
             }
